backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.2.0
